+++ node/middlewares/checkout.ts
@@ -271,6 +271,8 @@ export async function digitalRiverCreateCheckout(
     taxIdResult,
   })
 
+  const dock = docks.length > 0 && docks[0]
+
   const checkoutPayload: DRCheckoutPayload = {
     applicationId,
     currency: orderFormData?.storePreferencesData?.currencyCode ?? 'USD',
@@ -281,12 +283,29 @@ export async function digitalRiverCreateCheckout(
     browserIp,
     shipFrom: {
       address: {
-        line1: orderFormData.shippingData?.address?.street || 'Unknown',
-        line2: orderFormData.shippingData?.address?.complement || '',
-        city: orderFormData.shippingData?.address?.city || 'Unknown',
-        state: orderFormData.shippingData?.address?.state || '',
-        postalCode: orderFormData.shippingData?.address?.postalCode || '',
-        country: shippingCountry,
+        line1:
+          dock?.address?.street ||
+          orderFormData.shippingData?.address?.street ||
+          'Unknown',
+        line2:
+          dock?.address?.complement ||
+          orderFormData.shippingData?.address?.complement ||
+          '',
+        city:
+          dock?.address?.city ||
+          orderFormData.shippingData?.address?.city ||
+          'Unknown',
+        state:
+          dock?.address?.state ||
+          orderFormData.shippingData?.address?.state ||
+          '',
+        postalCode:
+          dock?.address?.postalCode ||
+          orderFormData.shippingData?.address?.postalCode ||
+          '',
+        country: dock?.address?.country?.acronym
+          ? convertIso3To2(dock.address.country.acronym)
+          : shippingCountry,
       },
     },
     shipTo: {
@@ -487,19 +506,51 @@ export async function digitalRiverGetSources(
     : sessionData.profile
 
   const customerId = profileData?.id
+  const email = profileData?.email
   let customer
 
   if (customerId) {
     try {
-      customer = await digitalRiver.getCustomerById({ settings, customerId })
+      customer = await digitalRiver.getCustomerById({
+        settings,
+        customerId,
+      })
+    } catch (err) {
+      if (err.response.status !== 404) {
+        logger.error({
+          error: err,
+          message: 'DigitalRiverGetSources-getCustomerById',
+        })
+
+        throw new ResolverError({
+          message: 'Get Customer failure',
+          error: err,
+        })
+      }
+    }
+  }
+
+  if (!customer && customerId) {
+    try {
+      const customerPayload: DRCustomerPayload = {
+        id: customerId,
+        email,
+      }
+
+      customer = await digitalRiver.createCustomer({
+        settings,
+        customerPayload,
+      })
     } catch (err) {
       logger.error({
         error: err,
-        message: 'DigitalRiverGetSources-getCustomerById',
+        customerId,
+        email,
+        message: 'DigitalRiverCreateCheckout-createCustomerFailure',
       })
 
       throw new ResolverError({
-        message: 'Get Customer failure',
+        message: 'Create customer failed',
         error: err,
       })
     }